+++ testing/marionette/client/marionette/emulator.py
@@ -3,20 +3,25 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 import datetime
+from errors import *
+from mozdevice import devicemanagerADB, DMError
 from mozprocess import ProcessHandlerMixin
-import multiprocessing
 import os
 import re
 import platform
 import shutil
 import socket
 import subprocess
+import sys
 from telnetlib import Telnet
 import tempfile
 import time
+import traceback
 
 from emulator_battery import EmulatorBattery
 from emulator_geo import EmulatorGeo
+from emulator_screen import EmulatorScreen
+
 
 class LogcatProc(ProcessHandlerMixin):
     """Process handler for logcat which saves all output to a logfile.
@@ -39,8 +44,9 @@ class Emulator(object):
 
     def __init__(self, homedir=None, noWindow=False, logcat_dir=None,
                  arch="x86", emulatorBinary=None, res='480x800', sdcard=None,
-                 userdata=None, gecko_path=None):
+                 userdata=None):
         self.port = None
+        self.dm = None
         self._emulator_launched = False
         self.proc = None
         self.marionette_port = None
@@ -55,6 +61,7 @@ class Emulator(object):
         self.res = res
         self.battery = EmulatorBattery(self)
         self.geo = EmulatorGeo(self)
+        self.screen = EmulatorScreen(self)
         self.homedir = homedir
         self.sdcard = sdcard
         self.noWindow = noWindow
@@ -62,7 +69,6 @@ class Emulator(object):
             self.homedir = os.path.expanduser(homedir)
         self.dataImg = userdata
         self.copy_userdata = self.dataImg is None
-        self.gecko_path = gecko_path
 
     def _check_for_b2g(self):
         if self.homedir is None:
@@ -152,6 +158,20 @@ class Emulator(object):
         else:
             return self.port is not None
 
+    def check_for_crash(self):
+        """
+        Checks if the emulator has crashed or not.  Always returns False if
+        we've connected to an already-running emulator, since we can't track
+        the emulator's pid in that case.  Otherwise, returns True iff
+        self.proc is not None (meaning the emulator hasn't been explicitly
+        closed), and self.proc.poll() is also not None (meaning the emulator
+        process has terminated).
+        """
+        if (self._emulator_launched and self.proc is not None
+                                    and self.proc.poll() is not None):
+            return True
+        return False
+
     def create_sdcard(self, sdcard):
         self._tmp_sdcard = tempfile.mktemp(prefix='sdcard')
         sdargs = [self.mksdcard, "-l", "mySdCard", sdcard, self._tmp_sdcard]
@@ -231,7 +251,7 @@ class Emulator(object):
                 os.remove(self._tmp_sdcard)
                 self._tmp_sdcard = None
             return retcode
-        if self.logcat_proc:
+        if self.logcat_proc and self.logcat_proc.proc.poll() is None:
             self.logcat_proc.kill()
         return 0
 
@@ -264,6 +284,34 @@ class Emulator(object):
         else:
             self._adb_started = False
 
+    def wait_for_system_message(self, marionette):
+        marionette.start_session()
+        marionette.set_context(marionette.CONTEXT_CHROME)
+        marionette.set_script_timeout(45000)
+        # Telephony API's won't be available immediately upon emulator
+        # boot; we have to wait for the syste-message-listener-ready
+        # message before we'll be able to use them successfully.  See
+        # bug 792647.
+        print 'waiting for system-message-listener-ready...'
+        try:
+            marionette.execute_async_script("""
+waitFor(
+    function() { marionetteScriptFinished(true); },
+    function() { return isSystemMessageListenerReady(); }
+);
+            """)
+        except ScriptTimeoutException:
+            print 'timed out'
+            # We silently ignore the timeout if it occurs, since
+            # isSystemMessageListenerReady() isn't available on
+            # older emulators.  45s *should* be enough of a delay
+            # to allow telephony API's to work.
+            pass
+        print 'done'
+        marionette.set_context(marionette.CONTEXT_CONTENT)
+        marionette.delete_session()
+
+
     def connect(self):
         self._check_for_adb()
         self.start_adb()
@@ -277,7 +325,8 @@ class Emulator(object):
             online, offline = self._get_adb_devices()
         self.port = int(list(online)[0])
 
-        self.install_gecko()
+        self.dm = devicemanagerADB.DeviceManagerADB(adbPath=self.adb,
+                                                    deviceSerial='emulator-%d' % self.port)
 
     def start(self):
         self._check_for_b2g()
@@ -285,8 +334,7 @@ class Emulator(object):
 
         qemu_args = self.args[:]
         if self.copy_userdata:
-            # Make a copy of the userdata.img for this instance of the emulator
-            # to use.
+            # Make a copy of the userdata.img for this instance of the emulator to use.
             self._tmp_userdata = tempfile.mktemp(prefix='marionette')
             shutil.copyfile(self.dataImg, self._tmp_userdata)
             qemu_args[qemu_args.index('-data') + 1] = self._tmp_userdata
@@ -306,7 +354,14 @@ class Emulator(object):
             online, offline = self._get_adb_devices()
         self.port = int(list(online - original_online)[0])
         self._emulator_launched = True
+
+        self.dm = devicemanagerADB.DeviceManagerADB(adbPath=self.adb,
+                                                    deviceSerial='emulator-%d' % self.port)
+
+        # bug 802877
+        time.sleep(10)
         self.geo.set_default_location()
+        self.screen.initialize()
 
         if self.logcat_dir:
             self.save_logcat()
@@ -314,30 +369,55 @@ class Emulator(object):
         # setup DNS fix for networking
         self._run_adb(['shell', 'setprop', 'net.dns1', '10.0.2.3'])
 
-        self.install_gecko()
-
-    def _save_logcat_proc(self, filename, cmd):
-        self.logcat_proc = LogcatProc(filename, cmd)
-        self.logcat_proc.run()
-        self.logcat_proc.processOutput()
-        self.logcat_proc.waitForFinish()
-        self.logcat_proc = None
-
-    def install_gecko(self):
+    def install_gecko(self, gecko_path, marionette):
         """
         Install gecko into the emulator using adb push.  Restart b2g after the
         installation.
         """
-        if not self.gecko_path:
-            return
-        # need to remount so we can write to /system/b2g
-        self._run_adb(['remount'])
-        self._run_adb(['shell', 'stop', 'b2g'])
-        self._run_adb(['shell', 'rm', '-rf', '/system/b2g/*.so'])
-        print 'installing gecko binaries'
-        self._run_adb(['push', self.gecko_path, '/system/b2g'])
-        print 'restarting B2G'
-        self._run_adb(['shell', 'start', 'b2g'])
+        # See bug 800102.  We use this particular method of installing
+        # gecko in order to avoid an adb bug in which adb will sometimes
+        # hang indefinitely while copying large files to the system
+        # partition.
+        push_attempts = 10
+
+        print 'installing gecko binaries...'
+
+        try:
+            # need to remount so we can write to /system/b2g
+            self._run_adb(['remount'])
+            for root, dirs, files in os.walk(gecko_path):
+                for filename in files:
+                    rel_path = os.path.relpath(os.path.join(root, filename), gecko_path)
+                    system_b2g_file = os.path.join('/system/b2g', rel_path)
+                    for retry in range(1, push_attempts+1):
+                        print 'pushing', system_b2g_file, '(attempt %s of %s)' % (retry, push_attempts)
+                        try:
+                            self.dm.pushFile(os.path.join(root, filename), system_b2g_file)
+                            break
+                        except DMError:
+                            if retry == push_attempts:
+                                raise
+
+            print 'restarting B2G'
+            # see bug 809437 for the path that lead to this madness
+            self.dm.shellCheckOutput(['stop', 'b2g'])
+            time.sleep(10)
+            self.dm.shellCheckOutput(['start', 'b2g'])
+
+            if not self.wait_for_port():
+                raise TimeoutException("Timeout waiting for marionette on port '%s'" % self.marionette_port)
+            self.wait_for_system_message(marionette)
+
+        except (DMError, MarionetteException):
+            # Bug 812395 - raise a single exception type for these so we can
+            # explicitly catch them elsewhere.
+
+            # print exception, but hide from mozharness error detection
+            exc = traceback.format_exc()
+            exc = exc.replace('Traceback', '_traceback')
+            print exc
+
+            raise InstallGeckoError("unable to restart B2G after installing gecko")
 
     def rotate_log(self, srclog, index=1):
         """ Rotate a logfile, by recursively rotating logs further in the sequence,
@@ -359,12 +439,8 @@ class Emulator(object):
             self.rotate_log(filename)
         cmd = [self.adb, '-s', 'emulator-%d' % self.port, 'logcat']
 
-        # We do this in a separate process because we call mozprocess's
-        # waitForFinish method to process logcat's output, and this method
-        # blocks.
-        proc = multiprocessing.Process(target=self._save_logcat_proc, args=(filename, cmd))
-        proc.daemon = True
-        proc.start()
+        self.logcat_proc = LogcatProc(filename, cmd)
+        self.logcat_proc.run()
 
     def setup_port_forwarding(self, remote_port):
         """ Set up TCP port forwarding to the specified port on the device,
